requirements.txt
pathway[xpack-llm]>=0.0

openai>=1.0.0
python-dotenv>=1.0.0

pandas>=2.0.0
numpy>=1.24.0

nltk>=3.8.1
scikit-learn>=1.3.0
